util: virObjectLockable: convert virMutex to GMutex
Because virCond usage is connected to virMutex, this patch also converts
virCond to GCond.

An important note to make here: GMutex doesn't like when you try to
clear a lock that's still locked. So I had to hunt the few cases where
that happened and add an explicit virObjectUnlock call before unreffing
the object.

Message-Id: <20200410135510.2148319-43-r4f4rfs@gmail.com>

+++ src/qemu/qemu_hotplug.h
@@ -151,7 +151,7 @@ int qemuDomainSetVcpuInternal(virQEMUDriverPtr driver,
                               virBitmapPtr vcpus,
                               bool state);
 
-unsigned long long qemuDomainGetUnplugTimeout(virDomainObjPtr vm);
+gint64 qemuDomainGetUnplugTimeout(virDomainObjPtr vm);
 
 int qemuHotplugAttachDBusVMState(virQEMUDriverPtr driver,
                                  virDomainObjPtr vm,
